parser simplification (#186)
- Constrain MAX_PAYLOAD_LEN to the appropriate data type of u16
- Move Checksum to separate module and encapsulate checksum calculations more
- Improve documentation of Checksum calculator
- Make the Checksum calculator support streaming checksum calculations and a one shot calculation
- Add tests for the checksum calculator

+++ ublox/src/proto14.rs
@@ -19,7 +19,7 @@ pub struct Proto14;
 
 impl crate::UbxProtocol for Proto14 {
     type PacketRef<'a> = PacketRef<'a>;
-    const MAX_PAYLOAD_LEN: usize = packetref_proto14::MAX_PAYLOAD_LEN as usize;
+    const MAX_PAYLOAD_LEN: u16 = packetref_proto14::MAX_PAYLOAD_LEN;
 
     fn match_packet(
         class_id: u8,
